Add shape and type attributes to OutputPort
Optional shape and type attributes exist on InputPort already. It
seems these should be added to OutputPort as well. This is
particularly useful when generating ONNX models from MDF because
an output shape for the model is required always it seems. These
attributes on both input and output ports will be needed to
work on a shape inference system for MDF.