Updated SimpleExample.ipynb
@pgleeson @sanjayankur31
I've added a node to the previous graph and tried out changing parameters and running the graph. This is what I did in the past 2-3 days, I'm still making changes in the notebook that I'll upload in my next commit. Please suggest me some changes that I should make in the notebook.