Add a fixture to pytorch test export to add repo root to sys path.
These tests import from examples, but repo root is not always on
sys path (when pytest command is used rather that python -m pytest).
This should fix the issue for now but it might be better to not have
tests import from examples directly. We already test the examples.
Maybe we should just move this logic and asserts to the examples themselves
in this case.