Arrays PyTorch example was broken on new PyTorch.
This fix is a band-aid. The root issue is that the ONNX export from torch
removes unused graph inputs during optimization passes. Then when the
example passes and input to when running the ONNX graph things fail.
The graph will run without the input passed but this is probably
confusing as an example.