Made easier to fetch output nodes and values from EvaluableGraph.
The PyTorch example were having some issues where the output node name was
non-deterministic. This was makeing the test case fail because the output
node name was looked up with a hardcoded value. I added some helpers on EvaluableGraph.

- output_nodes: Similar to root_nodes but for the output nodes. This should probably be put on Graph instead but it is here for now.
- output_enodes: The same but a list of EvaluableNodes rather than nodes.
- get_output(id=None): When None is passed to EvaluableNode.get_output(). It
gets the outputs for all nodes. Returns a tuple if multiple.