execution_engine: Threshold: try to infer target shape if no curr_value
Before first execution, evaluable object curr_value is None. The
value it will take on after execution may be of any shape, which can
cause the Threshold comparison check to except. In case of Threshold
referencing an InputPort, OutputPort, or stateful Parameter, we can
determine the future shape. Otherwise, we don't know, and this remains
unfixed with non-scalar values.