Fix a bunch of issues for Pytorch->MDF for VGG16, ResNet, CNN.
- Properly handle ONNX operations that return multiple outputs. Now expressions for
the value of and output port cand index into the tuple, for example.
OutputPort(id='_13', value="onnx_BatchNormalization_1[0]"). This was needed for ResNet
This also fixed an existing hack for MaxPool. Might be more of these lying around, need
to take a better look.
- In order to get the above working, I changed "onnx::" to "onnx_" in function\parameter ids generated by the exported. This lets eval work because
"::" is a can't be parsed in Python.
- Moved modeic ONNX opset version to 15 from 13.
- Fixed issue where improper schema (wrong opset) was being looked up for
ops in the exporter now that ONNX has moved to 15.