Set tolerance on allclose to test_simple_convolution for now
I am pretty sure this is related to not handling precision consistently
in the execution engine. We are converting back and forth in different
places from float32 to float64. The results are matching intermittently
without a tolerance but I have set and absolute tolerance for now. Need
to investigate thie further.