Change CI workflow to install PsyNeuLink after core dependencies have been installed.
Currently, we are not testing if core mdf works without PyTorch or PsyNeuLink because
these are installed before core package is installed (PsyNeuLink depends on PyTorch).
I have moved this install to the end, after installation of all backends (including PNL) from PIP. This is better for testing that a clean install works.