Refactor tests for torchvision models.
@parikshit14, I changed some of your code to remove a lot of duplication.
test_import.py under pytorch now enumerates all models in torchvision and
creates a test for each different type of model. All the models have the
same interface so we can consolidate the testing code into a parameterized
pytest.

We have 5 out of 20 tests failing still, lets see if we can track down why
each of these models is failing to run in the execution engine.